Conflicts:
	drivers/dma/fsl-edma-common.c
	drivers/i2c/busses/i2c-imx-lpi2c.c
	drivers/media/platform/amphion/vpu_malone.c
	drivers/media/platform/amphion/vpu_v4l2.c
	drivers/media/platform/amphion/vpu_v4l2.h
	drivers/mmc/host/Kconfig
	drivers/net/ethernet/freescale/enetc/enetc.h
	drivers/net/phy/realtek/realtek_main.c
	kernel/sched/cpufreq_schedutil.c

+++ arch/arm64/include/asm/kvm_hypevents.h
@@ -99,9 +99,9 @@ HYP_EVENT(host_mem_abort,
 );
 
 HYP_EVENT(__hyp_printk,
-	HE_PROTO(u8 fmt_id, u64 a, u64 b, u64 c, u64 d),
+	HE_PROTO(u16 fmt_id, u64 a, u64 b, u64 c, u64 d),
 	HE_STRUCT(
-		he_field(u8, fmt_id)
+		he_field(u16, fmt_id)
 		he_field(u64, a)
 		he_field(u64, b)
 		he_field(u64, c)
